requirements.txt
tensorboard==2.16.2
scipy==1.13.1
opencv-python==4.9.0.80
tqdm==4.66.4
matplotlib==3.9.0
PyYAML
setuptools==59.5.0
pandas==2.2.1
scikit-learn==1.5.0
natsort==8.4.0
bokeh==2.4.3
torch==2.0.1
torchvision==0.15.2
segmentation-models-pytorch==0.3.3
xarray
loguru
albumentations
seaborn
xrft
